Change Float and Integers to Numeric

+++ spec/services/forecast_service_spec.rb
@@ -32,34 +32,34 @@ describe ForecastService, :vcr do
     expect(data[:current][:sunset]).to be_an Integer
 
     expect(data[:current]).to have_key(:temp)
-    expect(data[:current][:temp]).to be_a Float
+    expect(data[:current][:temp]).to be_a Numeric
 
     expect(data[:current]).to have_key(:feels_like)
-    expect(data[:current][:feels_like]).to be_a Float
+    expect(data[:current][:feels_like]).to be_a Numeric
 
     expect(data[:current]).to have_key(:pressure)
-    expect(data[:current][:pressure]).to be_an Integer
+    expect(data[:current][:pressure]).to be_an Numeric
 
     expect(data[:current]).to have_key(:humidity)
-    expect(data[:current][:humidity]).to be_an Integer
+    expect(data[:current][:humidity]).to be_an Numeric
 
     expect(data[:current]).to have_key(:dew_point)
-    expect(data[:current][:dew_point]).to be_a Float
+    expect(data[:current][:dew_point]).to be_a Numeric
 
     expect(data[:current]).to have_key(:uvi)
-    expect(data[:current][:uvi]).to be_a Float
+    expect(data[:current][:uvi]).to be_a Numeric
 
     expect(data[:current]).to have_key(:clouds)
-    expect(data[:current][:clouds]).to be_an Integer
+    expect(data[:current][:clouds]).to be_an Numeric
 
     expect(data[:current]).to have_key(:visibility)
-    expect(data[:current][:visibility]).to be_an Integer
+    expect(data[:current][:visibility]).to be_an Numeric
 
     expect(data[:current]).to have_key(:wind_speed)
-    expect(data[:current][:wind_speed]).to be_a Float
+    expect(data[:current][:wind_speed]).to be_a Numeric
 
     expect(data[:current]).to have_key(:wind_deg)
-    expect(data[:current][:wind_deg]).to be_an Integer
+    expect(data[:current][:wind_deg]).to be_an Numeric
 
     expect(data[:current]).to have_key(:weather)
     expect(data[:current][:weather]).to be_an Array
@@ -82,7 +82,7 @@ describe ForecastService, :vcr do
     expect(data[:minutely][0][:dt]).to be_an Integer
 
     expect(data[:minutely][0]).to have_key(:precipitation)
-    expect(data[:minutely][0][:precipitation]).to be_an Integer
+    expect(data[:minutely][0][:precipitation]).to be_an Numeric
 
     expect(data[:hourly]).to be_an Array
     expect(data[:hourly][0]).to be_a Hash
@@ -90,37 +90,37 @@ describe ForecastService, :vcr do
     expect(data[:hourly][0][:dt]).to be_an Integer
 
     expect(data[:hourly][0]).to have_key(:temp)
-    expect(data[:hourly][0][:temp]).to be_a Float
+    expect(data[:hourly][0][:temp]).to be_a Numeric
 
     expect(data[:hourly][0]).to have_key(:feels_like)
-    expect(data[:hourly][0][:feels_like]).to be_a Float
+    expect(data[:hourly][0][:feels_like]).to be_a Numeric
 
     expect(data[:hourly][0]).to have_key(:pressure)
-    expect(data[:hourly][0][:pressure]).to be_an Integer
+    expect(data[:hourly][0][:pressure]).to be_an Numeric
 
     expect(data[:hourly][0]).to have_key(:humidity)
-    expect(data[:hourly][0][:humidity]).to be_an Integer
+    expect(data[:hourly][0][:humidity]).to be_an Numeric
 
     expect(data[:hourly][0]).to have_key(:dew_point)
-    expect(data[:hourly][0][:dew_point]).to be_a Float
+    expect(data[:hourly][0][:dew_point]).to be_a Numeric
 
     expect(data[:hourly][0]).to have_key(:uvi)
-    expect(data[:hourly][0][:uvi]).to be_a Float
+    expect(data[:hourly][0][:uvi]).to be_a Numeric
 
     expect(data[:hourly][0]).to have_key(:clouds)
-    expect(data[:hourly][0][:clouds]).to be_an Integer
+    expect(data[:hourly][0][:clouds]).to be_an Numeric
 
     expect(data[:hourly][0]).to have_key(:visibility)
-    expect(data[:hourly][0][:visibility]).to be_an Integer
+    expect(data[:hourly][0][:visibility]).to be_an Numeric
 
     expect(data[:hourly][0]).to have_key(:wind_speed)
-    expect(data[:hourly][0][:wind_speed]).to be_a Float
+    expect(data[:hourly][0][:wind_speed]).to be_a Numeric
 
     expect(data[:hourly][0]).to have_key(:wind_deg)
-    expect(data[:hourly][0][:wind_deg]).to be_an Integer
+    expect(data[:hourly][0][:wind_deg]).to be_an Numeric
 
     expect(data[:hourly][0]).to have_key(:pop)
-    expect(data[:hourly][0][:pop]).to be_an Integer
+    expect(data[:hourly][0][:pop]).to be_an Numeric
 
     expect(data[:hourly][0]).to have_key(:weather)
     expect(data[:hourly][0][:weather]).to be_an Array
@@ -148,43 +148,43 @@ describe ForecastService, :vcr do
     expect(data[:daily][0]).to have_key(:temp)
     expect(data[:daily][0][:temp]).to be_a Hash
     expect(data[:daily][0][:temp]).to have_key(:day)
-    expect(data[:daily][0][:temp][:day]).to be_a Float
+    expect(data[:daily][0][:temp][:day]).to be_a Numeric
     expect(data[:daily][0][:temp]).to have_key(:min)
-    expect(data[:daily][0][:temp][:min]).to be_a Float
+    expect(data[:daily][0][:temp][:min]).to be_a Numeric
     expect(data[:daily][0][:temp]).to have_key(:max)
-    expect(data[:daily][0][:temp][:max]).to be_a Float
+    expect(data[:daily][0][:temp][:max]).to be_a Numeric
     expect(data[:daily][0][:temp]).to have_key(:night)
-    expect(data[:daily][0][:temp][:night]).to be_a Float
+    expect(data[:daily][0][:temp][:night]).to be_a Numeric
     expect(data[:daily][0][:temp]).to have_key(:eve)
-    expect(data[:daily][0][:temp][:eve]).to be_a Float
+    expect(data[:daily][0][:temp][:eve]).to be_a Numeric
     expect(data[:daily][0][:temp]).to have_key(:morn)
-    expect(data[:daily][0][:temp][:morn]).to be_a Float
+    expect(data[:daily][0][:temp][:morn]).to be_a Numeric
 
     expect(data[:daily][0]).to have_key(:feels_like)
     expect(data[:daily][0][:feels_like]).to be_a Hash
     expect(data[:daily][0][:feels_like]).to have_key(:day)
-    expect(data[:daily][0][:feels_like][:day]).to be_a Float
+    expect(data[:daily][0][:feels_like][:day]).to be_a Numeric
     expect(data[:daily][0][:feels_like]).to have_key(:night)
-    expect(data[:daily][0][:feels_like][:night]).to be_a Float
+    expect(data[:daily][0][:feels_like][:night]).to be_a Numeric
     expect(data[:daily][0][:feels_like]).to have_key(:eve)
-    expect(data[:daily][0][:feels_like][:eve]).to be_a Float
+    expect(data[:daily][0][:feels_like][:eve]).to be_a Numeric
     expect(data[:daily][0][:feels_like]).to have_key(:morn)
-    expect(data[:daily][0][:feels_like][:morn]).to be_a Float
+    expect(data[:daily][0][:feels_like][:morn]).to be_a Numeric
 
     expect(data[:daily][0]).to have_key(:pressure)
-    expect(data[:daily][0][:pressure]).to be_an Integer
+    expect(data[:daily][0][:pressure]).to be_an Numeric
 
     expect(data[:daily][0]).to have_key(:humidity)
-    expect(data[:daily][0][:humidity]).to be_an Integer
+    expect(data[:daily][0][:humidity]).to be_an Numeric
 
     expect(data[:daily][0]).to have_key(:dew_point)
-    expect(data[:daily][0][:dew_point]).to be_a Float
+    expect(data[:daily][0][:dew_point]).to be_a Numeric
 
     expect(data[:daily][0]).to have_key(:wind_speed)
-    expect(data[:daily][0][:wind_speed]).to be_a Float
+    expect(data[:daily][0][:wind_speed]).to be_a Numeric
 
     expect(data[:daily][0]).to have_key(:wind_deg)
-    expect(data[:daily][0][:wind_deg]).to be_an Integer
+    expect(data[:daily][0][:wind_deg]).to be_an Numeric
 
     expect(data[:daily][0]).to have_key(:weather)
     expect(data[:daily][0][:weather]).to be_an Array
@@ -199,12 +199,12 @@ describe ForecastService, :vcr do
     expect(data[:daily][0][:weather][0][:icon]).to be_a String
 
     expect(data[:daily][0]).to have_key(:clouds)
-    expect(data[:daily][0][:clouds]).to be_an Integer
+    expect(data[:daily][0][:clouds]).to be_an Numeric
 
     expect(data[:daily][0]).to have_key(:pop)
-    expect(data[:daily][0][:pop]).to be_a Float
+    expect(data[:daily][0][:pop]).to be_a Numeric
 
     expect(data[:daily][0]).to have_key(:uvi)
-    expect(data[:daily][0][:uvi]).to be_a Float
+    expect(data[:daily][0][:uvi]).to be_a Numeric
   end
 end
